Creating two separate classes that share the interface. One for the word table, the other for loading words from a file
