+++ src/Keywords.java
@@ -1,13 +1,10 @@
-import java.util.Random;
-
-public class Keywords {
-    private static final Random RANDOM = new Random();
-
-    public static final String[] WORDS = {
+public class Keywords implements Word{
+       public static final String[] WORDS = {
             "KSIĘŻYC", "PIASEK", "MORZE", "BAŁTYK", "WAKACJE", "WIRUS", "KWARANTANNA", "URODZINY",
             "LAMBORGINI", "MERCEDES", "PIGUŁA", "SZTUKA", "SIEKIERA", "WARAN", "PODŁOGA"
     };
 
+    @Override
     public String nextWordToFind() {
         return WORDS[RANDOM.nextInt(WORDS.length)];
     }
