+++ src/Word.java
@@ -1,0 +1,7 @@
+import java.util.Random;
+
+public interface Word {
+    Random RANDOM = new Random();
+
+    String nextWordToFind();
+}
